fix(core): widen CoreError::invalid_handle parameter from u32 to u64
All handle APIs use u64, but invalid_handle() accepted u32, causing
every call site to truncate with `as u32`. This would produce wrong
error messages for handles exceeding u32::MAX.

Changed the parameter to u64 and removed all 8 truncation casts across
stream.rs (5), session.rs (2), and server.rs (1).

Closes #35

+++ crates/iroh-http-core/src/server.rs
@@ -105,7 +105,7 @@ pub fn respond(
 
     let sender = handles
         .take_req_sender(req_handle)
-        .ok_or_else(|| CoreError::invalid_handle(req_handle as u32))?;
+        .ok_or_else(|| CoreError::invalid_handle(req_handle))?;
     sender
         .send(ResponseHeadEntry { status, headers })
         .map_err(|_| CoreError::internal("serve task dropped before respond"))
